+++ Changelog.md
@@ -1,5 +1,10 @@
 ## Ignition Transport 3.0.0
 
+1. Support for an optional MessageInfo parameter in the user callbacks for
+   receiving messages. This parameter provides some information about the
+   message received (e.g.: topic name).
+    * [Pull request 191](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/191)
+
 1. Added `Node::Publisher::HasConnections` function that can be used to
    check if a Publisher has subscribers.
     * [Pull request 190](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/190)